+++ Optimizations/LinearDataAccess/Miners/WaterMinerExecutor.cs
@@ -157,14 +157,11 @@ internal sealed class WaterMinerExecutor
             }
 
             int outputBeltId = planet.entityPool[miner.insertTarget].beltId;
-            if (outputBeltId <= 0)
+            if (!beltExecutor.TryOptimizedCargoPath(planet, outputBeltId, out OptimizedCargoPath? outputBelt))
             {
                 continue;
             }
-
             int outputBeltOffset = planet.cargoTraffic.beltPool[outputBeltId].pivotOnPath;
-            CargoPath outputCargoPath = planet.cargoTraffic.pathPool[planet.cargoTraffic.beltPool[outputBeltId].segPathId];
-            OptimizedCargoPath outputBelt = beltExecutor.GetOptimizedCargoPath(outputCargoPath);
 
             OptimizedItemId productId = subFactoryProductionRegisterBuilder.AddProduct(planet.planet.waterItemId);
             int networkIndex = planet.powerSystem.consumerPool[miner.pcId].networkId;